requirements.txt
appdirs==1.4.2
decorator==4.0.11
Django==1.10.6
django-bootstrap3==8.2.1
django-extensions==1.7.7
djangorestframework==3.6.2
inflect==0.2.5
ipython==5.3.0
ipython-genutils==0.1.0
packaging==16.8
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.13
ptyprocess==0.5.1
Pygments==2.2.0
pyparsing==2.1.10
silly==1.1.0
simplegeneric==0.8.1
six==1.10.0
traitlets==4.3.2
wcwidth==0.1.7
